module6-task1 (done)
