--- js/components/common/abstract-view.js
@@ -1,30 +1,0 @@
-import {getDom} from "../util";
-
-export default class AbstractView {
-  constructor() {
-    if (new.target === AbstractView) {
-      throw new Error(`Can't instantiate AbstractView, only concrete one`);
-    }
-  }
-
-  get template() {
-    throw new Error(`Template is required`);
-  }
-
-  get element() {
-    if (this._element) {
-      return this._element;
-    }
-    this._element = this.render();
-    this.bind(this._element);
-    return this._element;
-  }
-
-  render() {
-    return getDom(this.template);
-  }
-
-  bind() {
-    // bind handlers if required
-  }
-}
