--- js/game-logic/start-game.js
@@ -1,6 +1,0 @@
-// import {updateGame} from "./update-game";
-// import {startTimer} from "./timer";
-
-// export const startGame = (game) => {
-//   updateGame(game);
-// };
