+++ js/main.js
@@ -1,8 +1,4 @@
-// import {selectTemplate} from "./components/util";
-// import WelcomeView from "./components/welcome";
 import {currentGame} from "./data/initial-game";
 import Router from "./router";
 
-// const welcome = new WelcomeView(currentGame);
-// selectTemplate(welcome.element);
 Router.showWelcome(currentGame);
