+++ js/views/common/abstract.js
