+++ js/views/game/game-genre.js
@@ -1,5 +1,5 @@
 import GameView from "./game";
-import {getDom} from "../components/util";
+import {getDom} from "../../components/util";
 
 export default class GameGenreView extends GameView {
   constructor(game, level) {
